Move `pool` config structs to a separate file
