+++ roles/pool/src/lib/mod.rs
@@ -1,3 +1,4 @@
+pub mod config;
 pub mod error;
 pub mod mining_pool;
 pub mod status;
@@ -5,8 +6,9 @@ pub mod template_receiver;
 
 use async_channel::{bounded, unbounded};
 
+use config::Configuration;
 use error::PoolError;
-use mining_pool::{get_coinbase_output, Configuration, Pool};
+use mining_pool::{get_coinbase_output, Pool};
 use template_receiver::TemplateRx;
 use tracing::{error, info, warn};
 
@@ -31,16 +33,16 @@ impl PoolSv2 {
         let (s_message_recv_signal, r_message_recv_signal) = bounded(10);
         let coinbase_output_result = get_coinbase_output(&config);
         let coinbase_output_len = coinbase_output_result?.len() as u32;
-        let tp_authority_public_key = config.tp_authority_public_key;
+        let tp_authority_public_key = config.tp_authority_public_key();
         TemplateRx::connect(
-            config.tp_address.parse().unwrap(),
+            config.tp_address().parse().unwrap(),
             s_new_t,
             s_prev_hash,
             r_solution,
             r_message_recv_signal,
             status::Sender::Upstream(status_tx.clone()),
             coinbase_output_len,
-            tp_authority_public_key,
+            tp_authority_public_key.cloned(),
         )
         .await?;
         let pool = Pool::start(
@@ -109,7 +111,7 @@ mod tests {
 
     #[tokio::test]
     async fn pool_bad_coinbase_output() {
-        let invalid_coinbase_output = vec![mining_pool::CoinbaseOutput::new(
+        let invalid_coinbase_output = vec![config::CoinbaseOutput::new(
             "P2PK".to_string(),
             "wrong".to_string(),
         )];
@@ -130,7 +132,7 @@ mod tests {
                 return;
             }
         };
-        config.coinbase_outputs = invalid_coinbase_output;
+        config.set_coinbase_outputs(invalid_coinbase_output);
         let pool = PoolSv2::new(config);
         let result = pool.start().await;
         assert!(result.is_err());
