+++ roles/pool/src/main.rs
@@ -2,7 +2,7 @@
 
 mod lib;
 use ext_config::{Config, File, FileFormat};
-pub use lib::{mining_pool::Configuration, status, PoolSv2};
+pub use lib::{config, status, PoolSv2};
 use tracing::error;
 
 mod args {
@@ -83,11 +83,11 @@ async fn main() {
     let config_path = args.config_path.to_str().expect("Invalid config path");
 
     // Load config
-    let config: Configuration = match Config::builder()
+    let config: config::Configuration = match Config::builder()
         .add_source(File::new(config_path, FileFormat::Toml))
         .build()
     {
-        Ok(settings) => match settings.try_deserialize::<Configuration>() {
+        Ok(settings) => match settings.try_deserialize::<config::Configuration>() {
             Ok(c) => c,
             Err(e) => {
                 error!("Failed to deserialize config: {}", e);
